renaming some types / variables
