+++ src/kostentraeger/edifact/segments.ts
@@ -7,7 +7,7 @@
 
 import { KostentraegerFilenameElements } from "../filename/parser"
 import { 
-    KostentraegerAbrechnungscodeSchluessel,
+    KostentraegerSGBVAbrechnungscodeSchluessel,
     AnschriftartSchluessel, 
     BundeslandSchluessel, 
     DatenlieferungsartSchluessel, 
@@ -15,7 +15,7 @@ import {
     IKVerknuepfungsartSchluessel, 
     KVBezirkSchluessel, 
     LeistungserbringergruppeSchluessel, 
-    KostentraegerPflegeLeistungsartSchluessel, 
+    KostentraegerSGBXILeistungsartSchluessel, 
     UebermittlungsmediumParameterSchluessel, 
     UebermittlungsmediumSchluessel, 
     UebermittlungszeichensatzSchluessel, 
@@ -122,7 +122,7 @@ export type VKG = {
      * 
      *  A value of 00 (Sammelschlüssel) means that this link is valid for all health care services 
      *  provided. */
-    pflegeLeistungsartSchluessel?: KostentraegerPflegeLeistungsartSchluessel,
+    sgbxiLeistungsartSchluessel?: KostentraegerSGBXILeistungsartSchluessel,
     /** A.k.a Leistungserbringerart. Only defined if leistungserbringergruppeSchluessel = 5 (Sonstige)
      *  
      *  The group/kind of health care service provided. Some of the possible values are groups. For
@@ -130,7 +130,7 @@ export type VKG = {
      *  
      *  A value of 00 (Sammelschlüssel) means that this link is valid for all health care services 
      *  provided. */
-    abrechnungscodeSchluessel?: KostentraegerAbrechnungscodeSchluessel,
+    sgbvAbrechnungscodeSchluessel?: KostentraegerSGBVAbrechnungscodeSchluessel,
     /** Only defined if leistungserbringergruppeSchluessel = 5 (Sonstige)
      *  
      *  A 5-digit numeral whose first 2 digits form the TarifbereichSchluessel and its last 3 digits
